adversarial_playground/tests/requirements.txt
pytest==7.1.2
pytest-cov==3.0.0
coverage
numpy==1.22.4
unittest2
art
google-cloud-storage
pydantic
fastapi[all]
requests==2.28.1
pandas==1.4.3
kagglehub
matplotlib==3.5.2
scikit-learn==1.1.2
opencv-python
httpx
python-multipart
adversarial-robustness-toolbox
scipy==1.8.1
tensorflow==2.9.1
tensorflow-addons==0.17.1
tensorboardX
ultralytics
ipython
wandb
keras==2.9.0
